+++ src/Components/Pocket.tsx
@@ -1,4 +1,6 @@
 import PokemonPocket from "./PokemonPocket";
+import gifUrl from "../assets/pokeball.gif"
+import { useState } from "react";
 
 export default function Pocket({
   pocket,
@@ -9,9 +11,14 @@ export default function Pocket({
   fightState: string;
   onSelect: (id: number) => void;
 }) {
+
+  const [isVisible, setIsVisible] = useState(false);
+
   return (
-    <div className=" absolute top-0 left-0 z-10 text-black">
-      <div className="flex flex-row">
+    <div className=" absolute bottom-0 left-0 z-50 text-black">
+      <div className="flex flex-row items-center">
+        <img onClick={()=>{setIsVisible(!isVisible)}} src={gifUrl} alt="pokeball" className="size-[80px] ml-8 mb-4 cursor-poke-full " />
+        <div className={`flex flex-row items-center ${isVisible ? "visible" : "invisible"}`}>
         {pocket.map((id) => (
           <PokemonPocket
             key={id}
@@ -22,6 +29,7 @@ export default function Pocket({
             }}
           />
         ))}
+        </div>
       </div>
     </div>
   );
