+++ src/Components/PokemonPocket.tsx
@@ -15,7 +15,7 @@ export default function PokemonPocket({
 
   return (
     <div
-      className="w-20"
+      className="max-w-16 items-center"
       onMouseEnter={() => setIsVisible(true)}
       onMouseLeave={() => setIsVisible(false)}
       onClick={() => {
@@ -28,9 +28,9 @@ export default function PokemonPocket({
       {isLoading && <h1 className="load-info">Loading...</h1>}
       {isSuccess && (
         <>
-          <div className="card w-full h-1/2 p-2 cursor-poke-full">
+          <div className="card p-2 cursor-poke-full">
             <img
-              className="w-auto h-full"
+              className="h-10"
               src={
                 data.sprites.other.showdown.front_default
                   ? data.sprites.other.showdown.front_default
@@ -40,7 +40,7 @@ export default function PokemonPocket({
               }
               alt={data.name}
             />
-                      <p className={`m-2 text-base ${isVisible ? "visible" : "invisible"}`}>
+                      <p className={`m-0 text-base ${isVisible ? "visible" : "invisible"}`}>
             {data.name}
           </p>
           </div>
