+++ src/data/backgrounds.ts
@@ -40,3 +40,52 @@ export const gifUrl = [gifUrl1,
     gifUrl19,
     gifUrl20
 ]
+
+
+
+
+
+
+
+
+export const gifColor = [ "#325c96",
+    "#44b9c8",
+    "#4e79a3",
+    "#720084",
+    "#f34021",
+    "#5288b2",
+    "#2fa4ed",
+    "#060200",
+    "#dba746",
+    "#6b9ce0",
+    "#595d76",
+    "#281a6e",
+    "#780000",
+    "#29a2e0",
+    "#fe7970",
+    "#785c4f",
+    "#3f317a",
+    "#4699bb",
+    "#a3b1da",
+    "#527ca4",]
+export const gifColorBottom = [ "#030226",
+    "#22281f",
+    "#000d22",
+    "#1b1716",
+    "#5a3e40",
+    "#132110",
+    "#142417",
+    "#1a1300",
+    "#151514",
+    "#1d2136",
+    "#595d76",
+    "#220b00",
+    "#000000",
+    "#172e10",
+    "#000000",
+    "#412e3f",
+    "#1a1d1d",
+    "#1a1d1d",
+    "#313439",
+    "#060200",
+]
