+++ src/data/coordinates.ts
@@ -34,13 +34,13 @@ export const coordinates = [
   [1231, 1155],
   [1655, 1384],
   [1211, 1503],
-  [1889, 1022],
+  [1729, 832],
   [1324, 653],
   [1402, 901],
   [346, 1899],
   [1325, 1981],
   [1389, 1944],
-  [248, 1097],
+  [348, 1097],
   [1306, 988],
 ];
